==> Analysis/1. Analyzing League Data.ipynb <==
# Analyzing League Data

#Importing Libraries
import numpy as np #To handle Mathematical calculations
import matplotlib.pyplot as plt #To plot charts 
%matplotlib inline
import pandas as pd #TO import and manage datasets
import glob
import os

# path = r'C:\Projects\FYP\Datasets\Whoscored\Points-Table' # use your path
path = r'../Datasets/Whoscored/Points-Table' # use your path
folder = glob.glob(path + "/*.csv")

li = []

for file in folder:
#     df = pd.read_csv(filename, index_col=None, header=0)
    individualTeamDataframe = pd.read_csv(file)
    league = os.path.basename(file).split('-')[0]
    season = os.path.basename(file).split('-')[3]
    individualTeamDataframe['League'] = league
    individualTeamDataframe['Season'] = int(season)
    li.append(individualTeamDataframe)

# combinedDataframe = pd.concat(li, axis=0, ignore_index=True)
combinedTeamDataframe = pd.concat(li)

# dataset = pd.read_csv('C:\Projects\FYP\Datasets\Whoscored\Points-Table\LaLiga-Points-Table-2018-2019.csv')
# league = 'LaLiga'

combinedTeamDataframe.head()
# dataset['League'] = league
# dataset

combinedTeamDataframe.shape

combinedTeamDataframe.info()

laligaTeamDf = combinedTeamDataframe.loc[combinedTeamDataframe['League'] == 'LaLiga']
bundesligaTeamDf = combinedTeamDataframe.loc[combinedTeamDataframe['League'] == 'Bundesliga']
ligue1TeamDf = combinedTeamDataframe.loc[combinedTeamDataframe['League'] == 'Ligue1']
serieaTeamDf = combinedTeamDataframe.loc[combinedTeamDataframe['League'] == 'SerieA']
eplTeamDf = combinedTeamDataframe.loc[combinedTeamDataframe['League'] == 'PremierLeague']

# laliga2010Df = combinedDataframe.loc[(combinedDataframe['League'] == 'LaLiga') & (combinedDataframe['Season'] == '2010')]
# laliga2018Df = combinedDataframe.loc[(combinedDataframe['League'] == 'LaLiga') & (combinedDataframe['Season'] == '2018')]

laligaTeamDf
