==> Analysis/3. Data Preprocessing.ipynb <==
#Importing Libraries

import numpy as np #To handle Mathematical calculations
import matplotlib.pyplot as plt #To plot charts 
import pandas as pd #To import and manage datasets
import glob
import os
import warnings
# pd.set_option('display.max_columns', None)  
# pd.set_option('display.max_rows', None)  

#Disable warning messages
warnings.simplefilter(action='ignore', category=FutureWarning)

#Combining All Offensive Data
path = r'../Datasets/Whoscored/Player-Stats'
folder = glob.glob(path + "/*Offensive*.csv")

li = []

for file in folder:
    individualOffensivePlayerDataframe = pd.read_csv(file)
    league = os.path.basename(file).split('-')[0]
    season = os.path.basename(file).split('-')[2]
    individualOffensivePlayerDataframe['League'] = league
    individualOffensivePlayerDataframe['Season'] = season
    for index, row in individualOffensivePlayerDataframe.iterrows():
        age = int(row["Age"]) - (2018-int(row["Season"]))
        individualOffensivePlayerDataframe.set_value(index, 'Age', age)
        
        #Cleaning Apps column
        if '(' in row["Apps"] :
            apps = int(row["Apps"].split("(")[0])
            individualOffensivePlayerDataframe.set_value(index, 'Apps', apps)
        else :
            apps = int(row["Apps"])
            individualOffensivePlayerDataframe.set_value(index, 'Apps', apps)
            
        #Cleaning Position column
        if ',' in row["Position"]:
            primary = row["Position"].split(",")[0]
            if '(' in primary:
                primary = primary.split("(")[0]
                individualOffensivePlayerDataframe.set_value(index, 'Position', primary)
            else:
                individualOffensivePlayerDataframe.set_value(index, 'Position', primary)
        elif '(' in row["Position"]:
            primary = row["Position"].split("(")[0]
            individualOffensivePlayerDataframe.set_value(index, 'Position', primary)
            
    li.append(individualOffensivePlayerDataframe)

combinedOffensivePlayerDataframe = pd.concat(li)

#Combining All Defensive Data
path = r'../Datasets/Whoscored/Player-Stats'
folder = glob.glob(path + "/*Defensive*.csv")

li = []

for file in folder:
    individualDefensivePlayerDataframe = pd.read_csv(file)
    league = os.path.basename(file).split('-')[0]
    season = os.path.basename(file).split('-')[2]
    individualDefensivePlayerDataframe['League'] = league
    individualDefensivePlayerDataframe['Season'] = season
    for index, row in individualDefensivePlayerDataframe.iterrows():
        age = int(row["Age"]) - (2018-int(row["Season"]))
        individualDefensivePlayerDataframe.set_value(index, 'Age', age)
        
        #Cleaning Apps column
        if '(' in row["Apps"] :
            apps = int(row["Apps"].split("(")[0])
            individualDefensivePlayerDataframe.set_value(index, 'Apps', apps)
        else :
            apps = int(row["Apps"])
            individualDefensivePlayerDataframe.set_value(index, 'Apps', apps)
            
        #Cleaning Position column  
        if ',' in row["Position"]:
            primary = row["Position"].split(",")[0]
            if '(' in primary:
                primary = primary.split("(")[0]
                individualDefensivePlayerDataframe.set_value(index, 'Position', primary)
            else:
                individualDefensivePlayerDataframe.set_value(index, 'Position', primary)
        elif '(' in row["Position"]:
            primary = row["Position"].split("(")[0]
            individualDefensivePlayerDataframe.set_value(index, 'Position', primary)
    
    li.append(individualDefensivePlayerDataframe)

combinedDefensivePlayerDataframe = pd.concat(li)

#Combining All Passing Data
path = r'../Datasets/Whoscored/Player-Stats'
folder = glob.glob(path + "/*Passing*.csv")

li = []

for file in folder:
    individualPassingPlayerDataframe = pd.read_csv(file)
    league = os.path.basename(file).split('-')[0]
    season = os.path.basename(file).split('-')[2]
    individualPassingPlayerDataframe['League'] = league
    individualPassingPlayerDataframe['Season'] = season
    for index, row in individualPassingPlayerDataframe.iterrows():
        age = int(row["Age"]) - (2018-int(row["Season"]))
        individualPassingPlayerDataframe.set_value(index, 'Age', age)
        
        #Cleaning Apps column
        if '(' in row["Apps"] :
            apps = int(row["Apps"].split("(")[0])
            individualPassingPlayerDataframe.set_value(index, 'Apps', apps)
        else :
            apps = int(row["Apps"])
            individualPassingPlayerDataframe.set_value(index, 'Apps', apps)
           
        #Cleaning Position column
        if ',' in row["Position"]:
            primary = row["Position"].split(",")[0]
            if '(' in primary:
                primary = primary.split("(")[0]
                individualPassingPlayerDataframe.set_value(index, 'Position', primary)
            else:
                individualPassingPlayerDataframe.set_value(index, 'Position', primary)
        elif '(' in row["Position"]:
            primary = row["Position"].split("(")[0]
            individualPassingPlayerDataframe.set_value(index, 'Position', primary)
    
    li.append(individualPassingPlayerDataframe)

combinedPassingPlayerDataframe = pd.concat(li)

#Combining All Summary Data
path = r'../Datasets/Whoscored/Player-Stats'
folder = glob.glob(path + "/*Summary*.csv")

li = []

for file in folder:
    individualSummaryPlayerDataframe = pd.read_csv(file)
    league = os.path.basename(file).split('-')[0]
    season = os.path.basename(file).split('-')[2]
    individualSummaryPlayerDataframe['League'] = league
    individualSummaryPlayerDataframe['Season'] = season
    for index, row in individualSummaryPlayerDataframe.iterrows():
        age = int(row["Age"]) - (2018-int(row["Season"]))
        individualSummaryPlayerDataframe.set_value(index, 'Age', age)
        
        #Cleaning Apps column
        if '(' in row["Apps"] :
            apps = int(row["Apps"].split("(")[0])
            individualSummaryPlayerDataframe.set_value(index, 'Apps', apps)
        else :
            apps = int(row["Apps"])
            individualSummaryPlayerDataframe.set_value(index, 'Apps', apps)
            
        #Cleaning Position column    
        if ',' in row["Position"]:
            primary = row["Position"].split(",")[0]
            if '(' in primary:
                primary = primary.split("(")[0]
                individualSummaryPlayerDataframe.set_value(index, 'Position', primary)
            else:
                individualSummaryPlayerDataframe.set_value(index, 'Position', primary)
        elif '(' in row["Position"]:
            primary = row["Position"].split("(")[0]
            individualSummaryPlayerDataframe.set_value(index, 'Position', primary)
            
    
    li.append(individualSummaryPlayerDataframe)

combinedSummaryPlayerDataframe = pd.concat(li)

#combining all the DFs'
combinedPlayerDataframe = pd.concat([combinedOffensivePlayerDataframe, combinedDefensivePlayerDataframe, combinedPassingPlayerDataframe, combinedSummaryPlayerDataframe], axis=1)

#Removing suplicates
combinedPlayerDataframe = combinedPlayerDataframe.loc[:,~combinedPlayerDataframe.columns.duplicated()]

#Removing players who have played less than 10 games in last season
minimum_number_of_games_played = 10
combinedPlayerDataframe = combinedPlayerDataframe[combinedPlayerDataframe.Apps >= minimum_number_of_games_played]

#Replace '-' with 0
combinedPlayerDataframe = combinedPlayerDataframe.replace('-', 0)

combinedPlayerDataframe

combinedPlayerDataframe.Position.unique()

combinedPlayerDataframe['Position'].replace(['M'], 'Midfielder', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['midfielder'], 'Midfielder', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['FW'], 'Forward', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['forward'], 'Forward', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['AM'], 'Attacking Midfielder', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['D'], 'Defender', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['GK'], 'Goalkeeper', regex=False, inplace=True)
combinedPlayerDataframe['Position'].replace(['DMC'], 'Defensive Midfielder', regex=False, inplace=True)

#write the DF to a csv file
combinedPlayerDataframe.to_csv('../Pre_Processed_Datasets/Preprocessed Player Data.csv', index = False)

#Combine all the league tables between 2018-2019

path = r'../Datasets/Whoscored/Points-Table'
folder = glob.glob(path + "/*2018-2019.csv")

li = []

for file in folder:
    individualLeagueTableDf = pd.read_csv(file)
    league = os.path.basename(file).split('-')[0]
    individualLeagueTableDf['League'] = league
    li.append(individualLeagueTableDf)
    
combinedLeagueTableDf = pd.concat(li)

#Keep only the League and Team columns

columns = combinedLeagueTableDf.columns.tolist()
columns.remove('Team')
columns.remove('League')
combinedLeagueTableDf.drop(columns, inplace=True, axis=1)

#Remove Eredivisie (it's not one of the top 5 leagues)
combinedLeagueTableDf = combinedLeagueTableDf[combinedLeagueTableDf.League != 'Eredivisie']

combinedLeagueTableDf

#write the DF to a csv file
combinedLeagueTableDf.to_csv('../Pre_Processed_Datasets/League-Team.csv', index = False)
