==> Analysis/4. Forecasting Active Player Ratings.ipynb <==
#Importing Libraries

import numpy as np #To handle Mathematical calculations
import matplotlib.pyplot as plt #To plot charts 
%matplotlib inline
import pandas as pd #TO import and manage datasets
import glob
import os
import warnings
# pd.set_option('display.max_columns', None)  
# pd.set_option('display.max_rows', None)  

#Disable warning messages
warnings.simplefilter(action='ignore', category=FutureWarning)

#Load preprocessed dataset
combinedPlayerDataframe = pd.read_csv("../Pre_Processed_Datasets/Preprocessed Player Data.csv") 

combinedPlayerDataframe.head()

#Selecting only the needed columns for the forecasting process

allPlayers = combinedPlayerDataframe[['Name', 'Season', 'Rating', 'Age']]
allPlayers = allPlayers[allPlayers['Season'].notnull()].copy()
allPlayers['Season'] = allPlayers['Season'].astype(int)
allPlayers

#Store a list of all the players
nameList = allPlayers.Name.unique()

#create a dataframe just to store the mean values of player rating's 
meanDf = pd.DataFrame(columns=['Mean'])

#Calculating career growth and rating avergae of every player (Feature engineering)
for name in nameList:
    mean = allPlayers.loc[allPlayers.Name == name]['Rating'].mean()
    length = len(allPlayers.loc[allPlayers.Name == name]['Rating'])
    growth = allPlayers.loc[allPlayers.Name == name]['Rating'].tolist()[length-1] - allPlayers.loc[allPlayers.Name == name]['Rating'].tolist()[0]
    meanDf.set_value(name, 'Name', name)
    meanDf.set_value(name, 'Mean', mean)
    meanDf.set_value(name, 'Growth', growth)
    
#Merging all the data together   
allPlayers = pd.merge(allPlayers, meanDf, on=['Name'])

#Converting column type
allPlayers['Mean'] = allPlayers['Mean'].astype(np.float64)

allPlayers.loc[allPlayers.Name == 'Lionel Messi']

#Another portion of feature engineering

from pandas import Series
from pandas import DataFrame
from pandas import concat

allPlayers = allPlayers.sort_values(['Season', 'Name'])

#Creating 2 new features for every player by combining both the last season rating and the difference between that and the
#~previous season rating

transformedPlayerDf = allPlayers.copy()
transformedPlayerDf['Last_Season_Rating'] = transformedPlayerDf.groupby(['Name'])['Rating'].shift()
transformedPlayerDf['Last_Season_Diff'] = transformedPlayerDf.groupby('Name')['Last_Season_Rating'].transform(Series.diff)
transformedPlayerDf = transformedPlayerDf.dropna()
transformedPlayerDf.head()

#Forecasting using linear regression

from statistics import mean 
from sklearn import linear_model 

from sklearn.model_selection import cross_val_score
from sklearn.metrics import mean_squared_error
from sklearn.metrics import median_absolute_error
from sklearn import model_selection
from math import sqrt

accuracy_MSE = []
accuracy_RMSE = []
accuracy_MAE = []
accuracy_R2 = []

#For every season using previous season data to forecast current season rating
for season in range(2015,2019):
    
    training_set = transformedPlayerDf[transformedPlayerDf['Season'] == season-1]
    testing_set = transformedPlayerDf[transformedPlayerDf['Season'] == season]

    X_train = training_set.drop(['Rating', 'Name', 'Season'], axis=1)
    y_train = training_set['Rating'].values
    X_test = testing_set.drop(['Rating', 'Name', 'Season'], axis=1)
    y_test = testing_set['Rating'].values

    lin_reg = linear_model.LinearRegression(n_jobs=-1) 
    
    lin_reg.fit(X_train, y_train)

    #using k-folds cross validation for calculating forecast accuracy
    scores = cross_val_score(lin_reg, X_test, y_test, cv=10)
    accuracy_R2.append(scores.mean())
    
    y_pred = lin_reg.predict(X_test)
    mse = mean_squared_error(y_test, y_pred)
    rmse = sqrt(mse)
    mae = median_absolute_error(y_test, y_pred)
    
    accuracy_MSE.append(mse)
    accuracy_RMSE.append(rmse)
    accuracy_MAE.append(mae)
     
print('Average accuracy for forecasting player ratings between 2015-2018:')
print("MSE: " + "{:.2f}".format(mean(accuracy_MSE)))
print("RMSE: " + "{:.2f}".format(mean(accuracy_RMSE)))
print("MAE: " + "{:.2f}".format(mean(accuracy_MAE)))
print("R2: " + "{:.2f}".format(mean(accuracy_R2)))
# accuracy

#Forecasting player ratings between 2019-2022

for season in range(2017, 2021):
    
    previousPlayers = allPlayers.loc[allPlayers.Season == season].copy()
    currentPlayers = allPlayers.loc[allPlayers.Season == season+1].copy()
    futurePlayers = currentPlayers.copy()

    futurePlayers['Season'] = currentPlayers['Season'].values+1
    futurePlayers['Age'] = currentPlayers['Age'].values+1

    futurePlayers.drop(['Rating'], axis=1)

    li = []
    li.append(previousPlayers)
    li.append(currentPlayers)
    li.append(futurePlayers)

    combinedDf = pd.concat(li)
    
    combinedDf = combinedDf.sort_values(['Season', 'Name'])

    transformedPlayerDf = combinedDf.copy()
    transformedPlayerDf['Last_Season_Rating'] = transformedPlayerDf.groupby(['Name'])['Rating'].shift()
    transformedPlayerDf['Last_Season_Diff'] = transformedPlayerDf.groupby('Name')['Last_Season_Rating'].transform(Series.diff)
    transformedPlayerDf = transformedPlayerDf.dropna()
    
    transformedPlayerDf = transformedPlayerDf.loc[transformedPlayerDf['Season'] == season+2]
    
    currentSeasonDf = transformedPlayerDf.drop(['Rating', 'Name', 'Season'], axis=1)
    y_pred = lin_reg.predict(currentSeasonDf)

    #Attaching the predicted values to the original DF so that the cycle can continue 
    #~until the forecasting range comes to an end 
    ratings = ["{:.2f}".format(value) for value in y_pred.tolist()]
    transformedPlayerDf['Rating'] = ratings
    transformedPlayerDf = transformedPlayerDf[['Name', 'Season', 'Rating', 'Age', 'Mean', 'Growth']]
    allPlayers = allPlayers.append(transformedPlayerDf)
    allPlayers = allPlayers[allPlayers['Rating'].notnull()].copy()
    allPlayers['Rating'] = allPlayers['Rating'].astype(float)


allPlayers.loc[allPlayers.Name == 'Kylian Mbappé']

allPlayers.loc[(allPlayers.Season == 2018) & ((allPlayers.Age == 24) | (allPlayers.Age == 25))]

#write the DF to a csv file
allPlayers.to_csv('../Pre_Processed_Datasets/Forecasted Ratings 2019-2022.csv', index = False)

# saving the model
import pickle
with open('../Models/model_forecasting_lr.pkl','wb') as file:
    pickle.dump(lin_reg, file)
