==> Analysis/5. Identifying potential transfer targets.ipynb <==
#Importing Libraries

import numpy as np #To handle Mathematical calculations
import matplotlib.pyplot as plt #To plot charts 
%matplotlib inline
import pandas as pd #To import and manage datasets
import glob
import os
import warnings
pd.set_option('display.max_columns', None)  
# pd.set_option('display.max_rows', None)  

#Disable warning messages
warnings.simplefilter(action='ignore', category=FutureWarning)
pd.options.mode.chained_assignment = None  # default='warn'

#Load preprocessed dataset
combinedPlayerDataframe = pd.read_csv("../Pre_Processed_Datasets/Preprocessed Player Data.csv") 

combinedPlayerDataframe.head()

#Load forecasted rating dataset
forecastedPlayerDataframe = pd.read_csv("../Pre_Processed_Datasets/Forecasted Ratings 2019-2022.csv") 

forecastedPlayerDataframe.head()

def get_player_df(name):
    player_df  = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Name == name) & (combinedPlayerDataframe.Season == 2018)]
    return player_df

def get_current_squad_df(team):
    squad_df = combined_player_data_frame.loc[(combined_player_data_frame.Team == team) & (combined_player_data_frame.Season == 2018)]
    return squad_df

### Handling young player data (Age: 23-28, Season 2018)

#Creating a seperate DF for players who are between the age of 23-28 in 2018

youngPlayers2018Df = forecastedPlayerDataframe.loc[(forecastedPlayerDataframe.Season == 2018) 
                                                   & ((forecastedPlayerDataframe.Age > 22) 
                                                       & (forecastedPlayerDataframe.Age < 29))]

#Storing the names of the young players on to list

youngPlayerNames = youngPlayers2018Df.Name.tolist()

#Creating a seperate dictionary entry for every player
#Creating a new row of data for every player with the mean of all their stats
#Rating has been replaced by the forecasted rating average

young_players_ = {}

for name in youngPlayerNames:
    playerMeanRating = forecastedPlayerDataframe.loc[(forecastedPlayerDataframe.Name == name)]['Rating'].mean()
    playerDf = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Name == name)]
    playerDf = playerDf.sort_values(by='Age', ascending=True)
    playerDf.loc['average'] =  playerDf.mean()
    playerDf.loc['average', 'Name'] = name
    playerDf.loc['average', 'Team'] = playerDf['Team'][playerDf.index[-2]]
    playerDf.loc['average', 'Position'] = playerDf['Position'][playerDf.index[-2]]
    playerDf.loc['average', 'League'] = playerDf['League'][playerDf.index[-2]]
    playerDf.loc['average', 'Rating'] = playerMeanRating
    playerDf = playerDf.iloc[-1:]
    young_players_[name] = playerDf

young_players_['Leroy Sané']

#Grouping youg player data into seperate position based dictionary entries

postionYoungPlayers_ = {}

Midfielders = []
Forwards = []
Attacking_Midfielders = []
Defenders = []
Goalkeepers = []
Defensive_Midfielders = []

for name in young_players_:
    position = young_players_[name].Position.tolist()[0]
    if position == 'Midfielder':
        Midfielders.append(young_players_[name])
    if position == 'Forward':
        Forwards.append(young_players_[name])
    if position == 'Attacking Midfielder':
        Attacking_Midfielders.append(young_players_[name])
    if position == 'Defender':
        Defenders.append(young_players_[name])
    if position == 'Goalkeeper':
        Goalkeepers.append(young_players_[name])
    if position == 'Defensive Midfielder':
        Defensive_Midfielders.append(young_players_[name])
    
postionYoungPlayers_['Midfielder']  = pd.concat(Midfielders)
postionYoungPlayers_['Forward'] = pd.concat(Forwards)
postionYoungPlayers_['Attacking Midfielder'] = pd.concat(Attacking_Midfielders)
postionYoungPlayers_['Defender'] = pd.concat(Defenders)
postionYoungPlayers_['Goalkeeper'] = pd.concat(Goalkeepers)
postionYoungPlayers_['Defensive Midfielder'] = pd.concat(Defensive_Midfielders)

postionYoungPlayers_['Defensive Midfielder'].head()

### Predicting replacements for older players

#Choose club

club_name = 'Real Madrid'

#Creating a sperate DF for older players

olderPlayersDf = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Team == club_name) 
                                            & (combinedPlayerDataframe.Season == 2018) 
                                            & (combinedPlayerDataframe.Age > 32)]

#Creating a seperate dictionary to store the average stat data of the best years (24-29) of the older players

older_players_ = {}

for index, row in olderPlayersDf.iterrows():
    
    currentDf = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Name == row['Name']) 
                                            & (combinedPlayerDataframe.Age > 23)  
                                            & (combinedPlayerDataframe.Age < 30)]
    currentDf = currentDf.sort_values(by='Age', ascending=True)
    currentDf.loc['average'] =  currentDf.mean()
    currentDf.loc['average', 'Name'] = row['Name']
    currentDf.loc['average', 'Team'] = currentDf['Team'][currentDf.index[-2]]
    currentDf.loc['average', 'Position'] = currentDf['Position'][currentDf.index[-2]]
    currentDf.loc['average', 'League'] = currentDf['League'][currentDf.index[-2]]
    currentDf = currentDf.iloc[-1:]
    older_players_[row['Name']] = currentDf

#Recommending young replacements within the age of (23-28)

li = []

old_player_ = {}

for name in older_players_:
    position = older_players_[name].Position.tolist()[0]
    li.append(older_players_[name])
    youngDf = postionYoungPlayers_[position]
    youngDf = youngDf[youngDf.Team != club_name]
    li.append(youngDf)
    
    combinedDf = pd.concat(li)
    
    stats = combinedDf.copy(deep=True)
    stats.drop(['Name', 'Team', 'League', 'Position'], axis=1, inplace=True)
    
    stats_with_name = stats
    stats = stats.dropna()
    stats_with_name = stats_with_name.dropna()
    stats_with_name["Name"] = combinedDf["Name"]
    
    stats_with_name = stats_with_name.reset_index(drop=True)
    
    from sklearn.preprocessing import StandardScaler
    from sklearn.neighbors import NearestNeighbors
    from sklearn.exceptions import DataConversionWarning

    import warnings
    warnings.filterwarnings(action="ignore",category=DataConversionWarning)
    
    scaler = StandardScaler()
    
    X = scaler.fit_transform(stats)
    
    #Limiting the number of similar profiles(clusters) to be found to 3. (4 in total with the current player being one)
    knn = NearestNeighbors(n_neighbors=4, algorithm='ball_tree').fit(X)

    player_indexes = knn.kneighbors(X)[1]
    
    print("")
    print("Older Player: " + name)
    
    old_player_[name] = get_player_df(name)
    
    index = stats_with_name[stats_with_name["Name"]==name].index.tolist()[0]
    count = 1
    for i in player_indexes[index][1:]:
            print('Option ' + str(count) + ': ' + stats_with_name.iloc[i]["Name"])
            
            old_player_[name] = old_player_[name].append(get_player_df(stats_with_name.iloc[i]["Name"]))
            
            count += 1

    
    li=[]
# print(old_player_['Luka Modric'])

### Predicting replacements for active underperforming players

#Choose club

club_name = 'Real Madrid'

activeClubPlayers = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Team == club_name) 
                                            & (combinedPlayerDataframe.Season == 2018) 
                                            & (combinedPlayerDataframe.Age > 27)
                                            & (combinedPlayerDataframe.Age < 33)]

activePlayerNames = activeClubPlayers.Name.tolist()

active_players_ = {}

for name in activePlayerNames:
    previouseFormDf = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Name == name)
                                               & (combinedPlayerDataframe.Age > 23)
                                               & (combinedPlayerDataframe.Age < 28)]
    
    recentFormDf = combinedPlayerDataframe.loc[(combinedPlayerDataframe.Name == name)
                                               & (combinedPlayerDataframe.Age > 27)
                                               & (combinedPlayerDataframe.Age < 33)]
    
    previouseRatingAvg = previouseFormDf.Rating.mean()
    recentRatingAvg = recentFormDf.Rating.mean()
    decline = (previouseRatingAvg-recentRatingAvg)
    
    if (recentRatingAvg < 8) & (decline >= 0.3) :
#         print('Name: ' + name)
#         print('Previouse Avg: ' + "%.2f" % previouseRatingAvg)
#         print('Current Avg: ' + "%.2f" % recentRatingAvg)
#         print('decline: ' + "%.2f" % decline)
#         print("")
        
        previouseFormDf = previouseFormDf.sort_values(by='Age', ascending=True)
        previouseFormDf.loc['average'] =  previouseFormDf.mean()
        previouseFormDf.loc['average', 'Name'] = name
        previouseFormDf.loc['average', 'Team'] = previouseFormDf['Team'][previouseFormDf.index[-2]]
        previouseFormDf.loc['average', 'Position'] = previouseFormDf['Position'][previouseFormDf.index[-2]]
        previouseFormDf.loc['average', 'League'] = previouseFormDf['League'][previouseFormDf.index[-2]]
        previouseFormDf = previouseFormDf.iloc[-1:]
        active_players_[name] = previouseFormDf
    
    elif (recentRatingAvg < 7) & (decline >= 0.2):
#         print('Name: ' + name)
#         print('Previouse Avg: ' + "%.2f" % previouseRatingAvg)
#         print('Current Avg: ' + "%.2f" % recentRatingAvg)
#         print('decline: ' + "%.2f" % decline)
#         print("")
        
        previouseFormDf = previouseFormDf.sort_values(by='Age', ascending=True)
        previouseFormDf.loc['average'] =  previouseFormDf.mean()
        previouseFormDf.loc['average', 'Name'] = name
        previouseFormDf.loc['average', 'Team'] = previouseFormDf['Team'][previouseFormDf.index[-2]]
        previouseFormDf.loc['average', 'Position'] = previouseFormDf['Position'][previouseFormDf.index[-2]]
        previouseFormDf.loc['average', 'League'] = previouseFormDf['League'][previouseFormDf.index[-2]]
        previouseFormDf = previouseFormDf.iloc[-1:]
        active_players_[name] = previouseFormDf

#Recommending young replacements within the age of (23-28)

li = []

active_player_ = {}

for name in active_players_:
    position = active_players_[name].Position.tolist()[0]
    li.append(active_players_[name])
    youngDf = postionYoungPlayers_[position]
    youngDf = youngDf[youngDf.Team != club_name]
    li.append(youngDf)
    
    combinedDf = pd.concat(li)
    
    stats = combinedDf.copy(deep=True)
    stats.drop(['Name', 'Team', 'League', 'Position'], axis=1, inplace=True)
    
    stats_with_name = stats
    stats = stats.dropna()
    stats_with_name = stats_with_name.dropna()
    stats_with_name["Name"] = combinedDf["Name"]
    
    stats_with_name = stats_with_name.reset_index(drop=True)
    
    from sklearn.preprocessing import StandardScaler
    from sklearn.neighbors import NearestNeighbors
    from sklearn.exceptions import DataConversionWarning

    import warnings
    warnings.filterwarnings(action="ignore",category=DataConversionWarning)
    
    scaler = StandardScaler()
    
    X = scaler.fit_transform(stats)
    
    #Limiting the number of similar profiles(clusters) to be found to 3. (4 in total with the current player being one)
    knn = NearestNeighbors(n_neighbors=4, algorithm='ball_tree').fit(X)

    player_indexes = knn.kneighbors(X)[1]
    
    print("")
    print("Active Underperfoming Player: " + name)
    
    active_player_[name] = get_player_df(name)
    
    index = stats_with_name[stats_with_name["Name"]==name].index.tolist()[0]
    count = 1
    for i in player_indexes[index][1:]:
            print('Option ' + str(count) + ': ' + stats_with_name.iloc[i]["Name"])
            
            active_player_[name] = active_player_[name].append(get_player_df(stats_with_name.iloc[i]["Name"]))
            
            count += 1

    
    li=[]
# print(active_player_)

# saving the model
import pickle
with open('../Models/model_replacements_knn.pkl','wb') as file:
    pickle.dump(knn, file)
